+++ GameSaveSwapper/Main.cs
@@ -362,38 +362,44 @@ namespace GameSaveSwapper {
 
         private void deleteToolStripMenuItem_Click(object sender, EventArgs e) {
             var item = listView1.FocusedItem;
-            for (var i = 0; i < this.profiles.Count; i++) {
-                if (this.profiles[i].name.Equals(item.SubItems[0].Text)) {
-                    String profileLoc = this.profiles[i].storeLocation;
-                    if (IsDirectoryEmpty(profileLoc)) {
-                        Directory.Delete(this.profiles[i].storeLocation);
-                    } else {
-                        var result = MessageBox.Show(
-                            "There are files in the profile save location, do you wish to delete them? ",
-                            "Existing Files in Save Location", MessageBoxButtons.YesNoCancel, MessageBoxIcon.Warning);
-                        if (result == DialogResult.Yes) {
-                            try {
-                                Directory.Delete(this.profiles[i].storeLocation, true);
-                            } catch (IOException ex) {
-                                log.Error("ProfileDelete: Deleting folder w/ content failed: ",ex);
+            var deleteResult = MessageBox.Show("Are you sure you want to delete this profile?", "Confirm Deletion",
+                MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (deleteResult == DialogResult.Yes) {
+                for (var i = 0; i < this.profiles.Count; i++) {
+                    if (this.profiles[i].name.Equals(item.SubItems[0].Text)) {
+                        String profileLoc = this.profiles[i].storeLocation;
+                        if (IsDirectoryEmpty(profileLoc)) {
+                            Directory.Delete(this.profiles[i].storeLocation);
+                        } else {
+                            var result = MessageBox.Show(
+                                "There are files in the profile save location, do you wish to delete them? ",
+                                "Existing Files in Save Location", MessageBoxButtons.YesNoCancel,
+                                MessageBoxIcon.Warning);
+                            if (result == DialogResult.Yes) {
+                                try {
+                                    Directory.Delete(this.profiles[i].storeLocation, true);
+                                } catch (IOException ex) {
+                                    log.Error("ProfileDelete: Deleting folder w/ content failed: ", ex);
+                                }
+
+                            } else if (result == DialogResult.Cancel) {
+                                return;
                             }
-                            
-                        }else if (result == DialogResult.Cancel) {
-                            return;
                         }
+
+                        //maybe stop deletion ?
+                        this.profiles.RemoveAt(i);
                     }
-                    //maybe stop deletion ?
-                    this.profiles.RemoveAt(i);
                 }
+
+                /* Gets a list of profiles from global
+                 * Finds profile and removes it from list
+                 * Saves profiles to disk
+                 * Loads profiles from disk (should rewrite this)
+                 */
+                SaveProfiles(this.profiles);
+                LoadGroupsToList();
             }
-            
-            /* Gets a list of profiles from global
-             * Finds profile and removes it from list
-             * Saves profiles to disk
-             * Loads profiles from disk (should rewrite this)
-             */
-            SaveProfiles(this.profiles);
-            LoadGroupsToList();
         }
